Adding a way to completelly disable the Auth{N|Z}
To completely disable the auth completely do following:

* Use the @AllPermissive annotation in RestBase class (it's commented
* there)
* Set the DUMMY flag to true in SecurityIntegration class (it's
* commented there)
* Comment out the secure-deployment element for
* hawkular-inventory-dist.war in
* dist/src/main/resources/wildfly/patches/standalone.xsl in
* hawkular/hawkular repo
* Comment out the keycloak xml elements in jboss-web.xml web.xml (it's
* commented there)
