+++ hawkular-inventory-rest-api/src/main/java/org/hawkular/inventory/rest/json/RelationshipEmbeddedJacksonSerializer.java
@@ -21,7 +21,7 @@ import java.util.Map;
 
 import org.hawkular.inventory.api.model.CanonicalPath;
 import org.hawkular.inventory.api.model.Relationship;
-import org.hawkular.inventory.rest.Security;
+import org.hawkular.inventory.rest.security.EntityIdUtils;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
@@ -104,7 +104,7 @@ public class RelationshipEmbeddedJacksonSerializer extends JsonSerializer<Relati
         jsonGenerator.writeStartObject();
 
         jsonGenerator.writeFieldName(FIELD_ID);
-        jsonGenerator.writeString(Security.getStableId(entity));
+        jsonGenerator.writeString(EntityIdUtils.getStableId(entity));
 
         jsonGenerator.writeFieldName(FIELD_SHORT_ID);
         jsonGenerator.writeString(entity.getSegment().getElementId());
