+++ hawkular-inventory-rest-api/src/main/java/org/hawkular/inventory/rest/json/RelationshipJacksonDeserializer.java
@@ -33,7 +33,7 @@ import java.util.stream.StreamSupport;
 
 import org.hawkular.inventory.api.model.CanonicalPath;
 import org.hawkular.inventory.api.model.Relationship;
-import org.hawkular.inventory.rest.Security;
+import org.hawkular.inventory.rest.security.EntityIdUtils;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
@@ -93,7 +93,7 @@ public class RelationshipJacksonDeserializer extends JsonDeserializer<Relationsh
     }
 
     private void validatePath(String path){
-        if (!Security.isValidRestPath(path)) {
+        if (!EntityIdUtils.isValidRestPath(path)) {
             throw new IllegalArgumentException("Error during relationship deserialization," +
                     " unable to recognize following path: " + path);
         }
