+++ hawkular-inventory-rest-api/src/main/java/org/hawkular/inventory/rest/security/TenantId.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.inventory.rest;
+package org.hawkular.inventory.rest.security;
 
 /**
  * @author Lukas Krejci
