+++ hawkular-inventory-rest-api/src/main/java/org/hawkular/inventory/rest/security/dummy/PermissiveSecurity.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright 2015 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.hawkular.inventory.rest.security.dummy;
+
+import javax.inject.Singleton;
+
+import org.hawkular.inventory.api.model.CanonicalPath;
+import org.hawkular.inventory.rest.security.Security;
+
+/**
+ * Dummy permissive implementation.
+ *
+ * @author Jirka Kremser
+ * @since 0.3.4
+ */
+@Singleton
+@AllPermissive
+public class PermissiveSecurity implements Security {
+
+    @Override public Security.CreatePermissionCheckerFinisher canCreate(Class<?> entityType) {
+        return path -> true;
+    }
+
+    @Override public boolean canUpdate(CanonicalPath path) {
+        return true;
+    }
+
+    @Override public boolean canDelete(CanonicalPath path) {
+        return true;
+    }
+
+    @Override public boolean canAssociateFrom(CanonicalPath path) {
+        return true;
+    }
+
+    @Override public boolean canCopyEnvironment(CanonicalPath path) {
+        return true;
+    }
+}
