+++ hawkular-inventory-rest-api/src/main/webapp/WEB-INF/web.xml
@@ -31,6 +31,7 @@
     </param-value>
   </context-param>
 
+  <!-- to disable the auth, comment out the following elements -->
   <security-constraint>
     <web-resource-collection>
       <web-resource-name>REST endpoints</web-resource-name>
